Pr/rescue rangers/21

+++ Helper/FileHelper.cs
@@ -1,4 +1,6 @@
 ﻿using CarCareTracker.Models;
+using CsvHelper;
+using System.Globalization;
 using System.IO.Compression;
 
 namespace CarCareTracker.Helper
@@ -7,6 +9,7 @@ namespace CarCareTracker.Helper
     {
         string GetFullFilePath(string currentFilePath, bool mustExist = true);
         byte[] GetFileBytes(string fullFilePath, bool deleteFile = false);
+        string GetFileText(string fullFilePath);
         string MoveFileFromTemp(string currentFilePath, string newFolder);
         bool RenameFile(string currentFilePath, string newName);
         bool DeleteFile(string currentFilePath);
@@ -99,6 +102,15 @@ namespace CarCareTracker.Helper
             }
             return Array.Empty<byte>();
         }
+        public string GetFileText(string fullFilePath)
+        {
+            if (File.Exists(fullFilePath))
+            {
+                var fileText = File.ReadAllText(fullFilePath);
+                return fileText;
+            }
+            return string.Empty;
+        }
         public bool RestoreBackup(string fileName, bool clearExisting = false)
         {
             var fullFilePath = GetFullFilePath(fileName);
@@ -120,6 +132,7 @@ namespace CarCareTracker.Helper
                 var dataPath = Path.Combine(tempPath, StaticHelper.DbName);
                 var widgetPath = Path.Combine(tempPath, StaticHelper.AdditionalWidgetsPath);
                 var configPath = Path.Combine(tempPath, StaticHelper.LegacyUserConfigPath);
+                var serverConfigPath = Path.Combine(tempPath, StaticHelper.LegacyServerConfigPath);
                 if (Directory.Exists(imagePath))
                 {
                     var existingPath = Path.Combine(_webEnv.ContentRootPath, "data", "images");
@@ -206,6 +219,15 @@ namespace CarCareTracker.Helper
                     }
                     File.Move(configPath, StaticHelper.UserConfigPath, true);
                 }
+                if (File.Exists(serverConfigPath))
+                {
+                    //check if config folder exists.
+                    if (!Directory.Exists("data/config"))
+                    {
+                        Directory.CreateDirectory("data/config");
+                    }
+                    File.Move(serverConfigPath, StaticHelper.ServerConfigPath, true);
+                }
                 return true;
             }
             catch (Exception ex)
@@ -221,14 +243,41 @@ namespace CarCareTracker.Helper
             if (!Directory.Exists(tempPath))
                 Directory.CreateDirectory(tempPath);
             int fileIndex = 0;
+            List<AttachmentExportModel> urlAttachments = new List<AttachmentExportModel>();
             foreach (GenericReportModel reportModel in exportData)
             {
                 foreach (UploadedFiles file in reportModel.Files)
                 {
                     var fileToCopy = GetFullFilePath(file.Location);
                     var destFileName = $"{tempPath}/{fileIndex}_{reportModel.DataType}_{reportModel.Date.ToString("yyyy-MM-dd")}_{file.Name}{Path.GetExtension(file.Location)}";
-                    File.Copy(fileToCopy, destFileName);
-                    fileIndex++;
+                    if (File.Exists(fileToCopy))
+                    {
+                        File.Copy(fileToCopy, destFileName);
+                        fileIndex++;
+                    } else
+                    {
+                        //file not found, must be a URL
+                        urlAttachments.Add(new AttachmentExportModel { 
+                            DataType = reportModel.DataType.ToString(),
+                            Date = reportModel.Date.ToString("yyyy-MM-dd"),
+                            Name = file.Name,
+                            Location = file.Location
+                        });
+                    }
+                }
+            }
+            if (urlAttachments.Any())
+            {
+                //write csv file detailing all urls.
+                var destFileName = $"{tempPath}/link_attachments.csv";
+                using (var writer = new StreamWriter(destFileName))
+                {
+                    using (var csv = new CsvWriter(writer, CultureInfo.InvariantCulture))
+                    {
+                        //custom writer
+                        StaticHelper.WriteAttachmentExportModel(csv, urlAttachments);
+                    }
+                    writer.Dispose();
                 }
             }
             var destFilePath = $"{tempPath}.zip";
@@ -248,6 +297,7 @@ namespace CarCareTracker.Helper
             var dataPath = StaticHelper.DbName;
             var widgetPath = StaticHelper.AdditionalWidgetsPath;
             var configPath = StaticHelper.UserConfigPath;
+            var serverConfigPath = StaticHelper.ServerConfigPath;
             if (!Directory.Exists(tempPath))
                 Directory.CreateDirectory(tempPath);
             if (Directory.Exists(imagePath))
@@ -297,6 +347,12 @@ namespace CarCareTracker.Helper
                 var newPath = Path.Combine(tempPath, "config");
                 Directory.CreateDirectory(newPath);
                 File.Copy(configPath, $"{newPath}/{Path.GetFileName(configPath)}");
+            }
+            if (File.Exists(serverConfigPath))
+            {
+                var newPath = Path.Combine(tempPath, "config");
+                Directory.CreateDirectory(newPath);
+                File.Copy(serverConfigPath, $"{newPath}/{Path.GetFileName(serverConfigPath)}");
             }
             var destFilePath = $"{tempPath}.zip";
             ZipFile.CreateFromDirectory(tempPath, destFilePath);
